Update Main.java
Add correct code

+++ HW1/Задание 7/Main.java
@@ -1,11 +1,11 @@
 public class Main{
     public static void main(String[] args) {
-	int a=4;
-	int b=6;
-	int r=5;
+	double a = 4;
+	double b = 6;
+	double r = 5;
 
  	
-	if ( a*a + b*b > r*r )
+	if ( a * a + b * b > r * r )
 		System.out.println ("possible");
 	else
 	System.out.println ("unpossible");
